+++ src/seeders/init.seed.ts
@@ -226,7 +226,6 @@ type VistaData = {
     { nombreVista: "Restablecer Password Usuario", vistaSuperior: "Usuarios" },
     { nombreVista: "Editar Vistas Usuario", vistaSuperior: "Usuarios" },
 
-
     // { nombreVista: "Cuentas Corrientes" },
     { nombreVista: "Estado de Cuenta", vistaSuperior: "Cuentas Corrientes" },
     { nombreVista: "Liquidaciones", vistaSuperior: "Cuentas Corrientes" },
@@ -248,7 +247,6 @@ type VistaData = {
     { nombreVista: "Territorialidad", vistaSuperior: "Repertorio" },
     { nombreVista: "Ver Titularidad", vistaSuperior: "Repertorio" },
 
-
     // { nombreVista: "Productoras" },
     { nombreVista: "Buscar Productora", vistaSuperior: "Productoras" },
     { nombreVista: "Premios Gardel", vistaSuperior: "Productoras" },
@@ -264,7 +262,6 @@ type VistaData = {
     { nombreVista: "Bloquear Usuario", vistaSuperior: "Usuarios" },
     { nombreVista: "Restablecer Password Usuario", vistaSuperior: "Usuarios" },
     { nombreVista: "Editar Vistas Usuario", vistaSuperior: "Usuarios" },
-
 
     // { nombreVista: "Cuentas Corrientes" },
     { nombreVista: "Estado de Cuenta", vistaSuperior: "Cuentas Corrientes" },
@@ -284,7 +281,6 @@ type VistaData = {
     { nombreVista: "Buscar Repertorio", vistaSuperior: "Repertorio" },
     { nombreVista: "Conflictos", vistaSuperior: "Repertorio" },
     { nombreVista: "Ver Titularidad", vistaSuperior: "Repertorio" },
-
 
     // { nombreVista: "Productoras" },
     { nombreVista: "Premios Gardel", vistaSuperior: "Productoras" },
@@ -308,12 +304,8 @@ type VistaData = {
     { nombreVista: "Conflictos", vistaSuperior: "Repertorio" },
     { nombreVista: "Ver Titularidad", vistaSuperior: "Repertorio" },
 
-
     // { nombreVista: "Productoras" },
-    { nombreVista: "Premios Gardel", vistaSuperior: "Productoras" },
-
-    // { nombreVista: "Usuarios" },
-    { nombreVista: "Buscar Usuario", vistaSuperior: "Usuarios" },
+    { nombreVista: "Premios Gardel", vistaSuperior: "Productoras" },    
 
     // { nombreVista: "Cuentas Corrientes" },
     { nombreVista: "Estado de Cuenta", vistaSuperior: "Cuentas Corrientes" },